+++ src/Plugin/PluginService.js
@@ -41,13 +41,18 @@ class PluginService {
     const pluginOptions = options.get();
     const loaderOptions = pluginOptions.loaderOptions || {};
 
-    // add the preprocessor option into the loader option
-    if (pluginOptions.preprocessor) {
+    // add reference for the preprocessor option into the loader options
+    if (pluginOptions.preprocessor && !loaderOptions.preprocessor) {
       loaderOptions.preprocessor = pluginOptions.preprocessor;
 
-      if (pluginOptions.preprocessorOptions) {
+      if (pluginOptions.preprocessorOptions && !loaderOptions.preprocessorOptions) {
         loaderOptions.preprocessorOptions = pluginOptions.preprocessorOptions;
       }
+    }
+
+    // add reference for data to the loader options
+    if (pluginOptions.data && !loaderOptions.data) {
+      loaderOptions.data = pluginOptions.data;
     }
 
     this.#used = true;