+++ test/cases/option-js-filename-undefined/webpack.config.js
@@ -1,0 +1,23 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+    filename: 'js/[name].[contenthash:8].js',
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: {
+        index: './src/index.html',
+      },
+      js: {
+        // must be used the output.filename
+        filename: undefined,
+      },
+    }),
+  ],
+};